Make the decision of whether or not to initiate a swap stochastic
This is how it is done in the van Steen & Tanenbaum book to ensure
that not every node initiates a swap at the same time. My
implementation maybe doesn't require it because I allow an incoming
swap request while already engaged in a swap with another
node. However, this does seem error prone and potentially not wise so
I'm moving in the direction of the code from the book.